# OpenApi/openAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Spring boot template",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/healthcheck": {
      "get": {
        "operationId": "Healthcheck",
        "tags": [
          "healthcheck"
        ],
        "description": "Healthcheck endpoint",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/provider": {
      "post": {
        "operationId": "addProvider",
        "tags": [
          "provider"
        ],
        "description": "Insert provider",
        "parameters": [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          }
        ],
        "requestBody": {
          "$ref" : "APIspecs/requests/providerPostRequestBody.json"
        },
        "responses": {
          "201": {
            "$ref": "APIspecs/responses/providerCreated.json"
          },
          "400": {
            "$ref": "APIspecs/responses/badRequest.json"
          },
          "409": {
            "$ref": "APIspecs/responses/conflict.json"
          }
        }
      },
      "get" : {
        "operationId": "getAllProviders",
        "tags": [
          "provider"
        ],
        "description": "Get all providers",
        "parameters": [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/queryParams/pageQueryParam.json"
          },
          {
            "$ref" : "common/params/queryParams/pageSizeQueryParam.json"
          }
        ],
        "responses" : {
          "200" : {
            "$ref" : "APIspecs/responses/providerGetAllOK.json"
          },
          "400": {
            "$ref": "APIspecs/responses/badRequest.json"
          }
        }        
      }
    },
    "/api/v1/provider/{provider-id}/item": {
      "post": {
        "operationId": "addItem",
        "tags": [
          "item"
        ],
        "description" : "Add item",
        "parameters": [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/providerIDParam.json"
          }
        ],
        "requestBody": {
          "$ref" : "APIspecs/requests/itemPostRequestBody.json"
        },
        "responses": {
          "201": {
            "$ref": "APIspecs/responses/itemCreated.json"
          },
          "400": {
            "$ref": "APIspecs/responses/badRequest.json"
          }
        }
      }
    },
    "/api/v1/provider/{provider-id}" : {
      "patch" : {
        "operationId": "patchProvider",
        "tags": [
          "provider"
        ],
        "description": "Update provider",
        "parameters": [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/providerIDParam.json"
          }
        ],
        "requestBody": {
          "$ref" : "APIspecs/requests/providerPatchRequestBody.json"
        },
        "responses": {
          "200": {
            "$ref": "APIspecs/responses/providerUpdated.json"
          },
          "400": {
            "$ref": "APIspecs/responses/badRequest.json"
          }
        }
      }
    },
    "/api/v1/provider/{provider-id}/item/{item-id}" : {
      "put": {
        "operationId": "updateItem",
        "tags" : [
          "item"
        ],
        "description" : "Update item",
        "parameters": [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/providerIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/itemIDParam.json"
          }
        ],
        "requestBody": {
          "$ref" : "APIspecs/requests/itemPutRequestBody.json"
        },
        "responses": {
          "200": {
            "$ref": "APIspecs/responses/itemUpdated.json"
          },
          "400": {
            "$ref": "APIspecs/responses/badRequest.json"
          }
        }
      },
      "get" : {
        "operationId": "getItemByID",
        "tags" : [
          "item"
        ],
        "description" : "Get item",
        "parameters": [
          {
            "$ref" : "common/params/headers/accountIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/providerIDParam.json"
          },
          {
            "$ref" : "common/params/pathParams/itemIDParam.json"
          }
        ],
        "responses": {
          "200": {
            "$ref": "APIspecs/responses/itemGetOK.json"
          },
          "404": {
            "$ref": "APIspecs/responses/notFound.json"
          }
        }        
      }
    }
  }
}